Pull out the simplify-segments code into a separate file.
This minimizes the amount of data that needs to be passed as input to
the plotmatplotlib step, which is the one that we'll want to iterate on
most often.
